requirements.txt
pandas>=1.3.0
numpy>=1.21.0,<2.0.0
scikit-surprise>=1.1.1
